requirements.txt
binance-connector=~2.0.0